requirements.txt
typing_extensions==4.13.2
requests==2.32.3
Pillow==9.5.0
keybert==0.9.0
tts==0.22.0
moviepy==1.0.3
duckduckgo-search==8.0.1
ollama==0.4.8
langchain==0.3.25
langgraph==0.4.3